# PERF: ArrowExtensionArray.to_numpy(dtype=object)