@jbrockmendel 2023-02-08
dumb question: why isn't `self._data.to_numpy()` good enough here?